+++ commands/toggle.js
@@ -1,4 +1,4 @@
-config = require("../config.json");
+configs = require("../server/config.json");
 const prefix = config.chars.prefix;
 const fs = require("fs");
 const bot = require("../bot.js");
@@ -9,7 +9,6 @@ module.exports = {
   aliases: ["toggle", "tog"],
   usage: `\`${prefix}toggle <setting>\``,
   args: true,
-  cooldown: 5,
 	guildOnly:true,
   permissions: "MANAGE_GUILD",
 	execute(message, args) {
@@ -18,30 +17,27 @@ module.exports = {
       return `, but it failed, as there were no arguments provided`;
     }
     try{
-      fs.readFile("config.json", (err, raw) => {
-        configs =JSON.parse(raw);
-        toggles = configs.toggles;
-        if (toggles[args[0]] === undefined){
-					message.lineReply(`Sorry, but ${args[0]} is not a valid setting. Use \`${prefix}show\` to see a list of all settings.`);
-          console.log(`But it failed, as ${args} is not a valid setting.`);
+      toggles = configs.toggles;
+      if (toggles[args[0]] === undefined){
+				message.lineReply(`Sorry, but ${args[0]} is not a valid setting. Use \`${prefix}show\` to see a list of all settings.`);
+        console.log(`But it failed, as ${args} is not a valid setting.`);
+        return;
+      }
+      was = toggles[args[0]];
+      toggles[args[0]] = !toggles[args[0]];
+      configs.toggles = toggles;
+      const jsonString = JSON.stringify(configs);
+      fs.writeFile("./server/config.json",jsonString, err => {
+        if (err) {
+          message.lineReply(`An unexpected error occured when editing the config file.`);
+          console.log(`But an unexpected write error occured. Error: ${err}`);
+          return;
+        } else {
+          message.lineReplyNoMention(`"${args}" was successfully toggled from \`${was.toString().replace("true","ON").replace("false","OFF")}\` to \`${toggles[args[0]].toString().replace("true","ON").replace("false","OFF")}\`.`);
+          console.log(`and successfully toggled "${args}" from ${was} to ${toggles[args[0]]}.`);
+					bot.loadConfigs();
           return;
         }
-        was = toggles[args[0]];
-        toggles[args[0]] = !toggles[args[0]];
-        configs.toggles = toggles;
-        const jsonString = JSON.stringify(configs);
-        fs.writeFile("./config.json",jsonString, err => {
-          if (err) {
-            message.lineReply(`An unexpected error occured when editing the config file.`);
-            console.log(`But an unexpected write error occured. Error: ${err}`);
-            return;
-          } else {
-            message.lineReplyNoMention(`"${args}" was successfully toggled from \`${was.toString().replace("true","ON").replace("false","OFF")}\` to \`${toggles[args[0]].toString().replace("true","ON").replace("false","OFF")}\`.`);
-            console.log(`and successfully toggled "${args}" from ${was} to ${toggles[args[0]]}.`);
-						bot.loadConfigs();
-            return;
-          }
-        })
       })
     } catch (err){
       message.lineReply(`An unexpected error occured.`);