+++ handlers/commands.js
@@ -1,8 +1,7 @@
-const prefix = require("../config.json").chars.prefix;
+const prefix = require("../server/config.json").chars.prefix;
 const {dateToTime} = require("../fun/dateToTime.js");
 
 function handleCommand(message,postedTime){
-  let time = dateToTime(postedTime);
   const client = message.client;
   if (!message.content.startsWith(prefix) || message.author.bot) return; //No prefix? Bot? Cancel
   //finangling the command and argument vars
@@ -13,7 +12,6 @@ function handleCommand(message,postedTime){
   //a bunch of checking
   if (!command) return; 																						//is it a command
   logString = `[${dateToTime(postedTime)}]: User ${message.author.username}${message.author} used ${prefix}${commandName}`;
-  console.log("testo 2");
   if (command.guildOnly && message.channel.type === "dm") { 				//dm checking
     logString = logString + `, but it failed, as ${prefix}${commandName} cannot be used in a DM`;
     console.log(logString);
@@ -35,24 +33,6 @@ function handleCommand(message,postedTime){
     logString = logString + `, but it failed, as it requires arguments, and none were provided.`;
     console.log(logString);
     return message.lineReply(reply);
-  }
-  if(command.cooldown){																							//per-author cooldown checking
-    if (!cooldowns.has(command.name)) {
-      cooldowns.set(command.name, new Discord.Collection());
-    }
-    const timestamps = cooldowns.get(command.name);
-    const cooldownAmount = (command.cooldown || 3) * 1000;
-    if (timestamps.has(message.author.id)) {
-      const expirationTime = timestamps.get(message.author.id) + cooldownAmount;
-      if (currentTime < expirationTime) {
-        const timeLeft = (expirationTime - currentTime) / 1000;
-        logString = logString + `, but it failed, as ${prefix}${commandName} was on cooldown from this user at the time.`;
-        console.log(logString);
-        return message.lineReply(`please wait ${timeLeft.toFixed(1)} more second(s) before reusing the \`${prefix}${command.name}\` command.`);
-      }
-    }
-    timestamps.set(message.author.id, currentTime);
-    setTimeout(() => timestamps.delete(message.author.id), cooldownAmount);
   }
   //command execution
   try {